Modification of POV support
Now QtGameControllerModif fully supports POV, that can be seen using the new app "Monitoring0". The RealJoystick class transforms the POV into 4 buttons, that can be seen in the app "Monitoring1". And in "Monitoring2" you can see the effect of the "ControllersPlugins" (naming of axes, buttons, reordering of axes, additional virtual buttons for some controllers, ...)