Led support for Warthog throttle
Done using the reimplementation of the RealJoystick::setData function in the ThrustmasterWarthogThrottle class (see code in ControllersPlugins/SOURCES). The new HidDevice class has been done using the Hid Device windows lib and with some help from https://www.edn.com/design/communications-networking/4336292/Using-the-HID-class-eases-the-job-of-writing-USB-device-drivers. To be able to compile it, you need to install the Windows lib. .. so I provided the binaries (dll) for the controllers plugins.