Star Citizen controls xml file added for the "3.1.4_v5" profile example
For the ones that would want to test it in the game. 2 vJoy devices must be configured with at least:
- 8 axes and 50 buttons for the first one
- 2 axes and 31 buttons for the second one
As I hide my TM warthog and MFG pedals with WhiteKnight, vJoy device 1 is "js1" and the second is "js2" for Star Citizen. You might encounter joystick indexing issues. It is a problem for everyone currently and CIG told they will change this (but not when and for what). In the meantime, the easier solution is to hide all your real joysticks with WhiteKnight or to do automatic replacement of "jsX" in the xml file so that it matches the right indexes for you.