Addition of "REXEC" feature
Allows to repeat an action or function at regular intervals until it is stopped. It is equivalent to Thrustmaster REXEC function with the RNOSTOP option.

Look at "startRexec" and "stopRexec" in AbstractProfile.h and in the example profile "Profile.cpp" to understand how it works... or in documentation when I will update it.