Bug fix: some events were skipped when using DoAction in an ActionCallback
The bug was introduced at the same time of "input repeater" in commit 37285883915960915f8d242ac99d706da88c9fdb. Here below is an example of impacted cases:
    auto swapYawRollToggle = [this]() {if (!tmwt->buttonPressed(TMWT::LDGH)) DoAction(new ActionButtonPulse{vj1,SC1::SwapYawRollToggle,ncPulse});};
    Map(tmwt, ControlType::Button, TMWT::APUON, AllLayers, new TriggerButtonChange{}, new ActionCallback{swapYawRollToggle});
When "APUON" button is pressed it generates an event of type "callback". This callback calls DoAction that generates new events that are added at the end of the events list. Among these events, the ones that were supposed to be sent immediatly are skipped... because of the classical problem of going through a variable size container. In the example above, no button pulse was noticed because the initial event (that presses the button down) was skipped.