Secondary apps: list controllers in another thread, not to freeze GUI.
When you connect or disconnect a controller, or modify the configuration of a vJoy device, if you want that the apps AxesCurves, Monitoring and ControllersInfo take this modification into account, you need to click on the "update" action in these apps. This update can take about 10s, mostly because of the time spent listing the controllers (calls to DirectInput and XInput).

Now these calls are done in another thread, so the GUI does not freeze anymore (no more "not responding"). During this time, you can not cancel the update or interact with the widget (a loading screen will be displayed). But you can quit the app or move the window.