Bugfix: in computation of profile dll path...
When starting the profile, an error said that the dll file of the profile was not found.
First the profiles are supposed to be in folders named as follows "ProfileName/release/xxxx.dll".
Secondly "UJPS.exe" was previously in a folder named "release" and is now in a folder named "bin".
Eventually UJPS main app computes the dll path as follows "QString dllDir = dirName(m_proFilePath) + "/" + debugOrRelease();" (MainWindow.cpp at line 429). So "debugOrRelease()" should return "release" and not "bin".